refactor: ConnectionManager 클래스 이동
