+++ nextstep-jdbc/build.gradle.kts
@@ -6,5 +6,5 @@ dependencies {
     testImplementation("org.springframework:spring-web:$springVersion")
 
     runtimeOnly("com.h2database:h2:1.4.199")
-    testImplementation("org.apache.commons:commons-dbcp2:2.6.0")
+    implementation("org.apache.commons:commons-dbcp2:2.6.0")
 }
