+++ nextstep-jdbc/src/main/java/nextstep/utils/QueryUtil.java
@@ -1,4 +1,4 @@
-package nextstep.util;
+package nextstep.utils;
 
 import java.util.Map;
